remove `supWhatsappLine` deep watch that reinitialize the form, loosing the previous selected messages.
This deepwatch was introduced at 611e05cdc by myself to fix OML-2861, but it is
causing problems now. Although I tend to avoid watches, this is how the current
UI is implemented.

Anyway looks like the deepwatch is not needed anymore since both the create/edit
form are reacting properly to changes in both the messages and destination
type dropdowns.